// navbar/src/app/models/global_set/company/company.ts
export class Company {
    companyId: string;
	parentId: number;
	companyType: string;
	companyName: string;
	companyGroup: string;
	establishSince: number;
	industryType: string;
	subindustryType: string;
	address: string;
	city: string;
	district: string;
	village: string;
	state: string;
	hamletNumber: string;
	neighbourhoodNumber: string;
	zipcode: number;
	companyLogo: string;
	updatedBy: number;
}